fix(ix-img): set directive attributes after view init
Currently, the directive's attributes are set during the `afterViewChecked` lifecycle hook. This can result in situations (such as adding an `onLoad` handler) where the directive can infinitely re-render. This commit changes the lifecycle hook where attributes are set to `afterViewInit` instead.

Perhaps there is a specific reason why `afterViewChecked` was originally used but to my understanding this hook shouldn't apply to ix-img since it does not have child components. However, this might not be the case for ix-source and/or ix-picture.

Resource: https://angular.io/guide/lifecycle-hooks#responding-to-view-changes